Revert "Add support for paged update_cache and fill_cache (#10523)"
This reverts commit 850216045b733160c15e9d0a975c8441373fdaf4.
