--- tests/ttnn/unit_tests/operations/test_paged_update_cache.py
@@ -1,635 +1,0 @@
-# SPDX-FileCopyrightText: © 2023 Tenstorrent Inc.
-
-# SPDX-License-Identifier: Apache-2.0
-
-import torch
-import pytest
-
-import tt_lib as ttl
-import ttnn
-from loguru import logger
-from models.utility_functions import nearest_32, pad_by_zero
-from tests.tt_eager.python_api_testing.sweep_tests.comparison_funcs import comp_pcc, comp_equal
-from models.utility_functions import is_grayskull
-
-
-def run_test_update_cache_decode(
-    cache_idx, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-):
-    input_shape = [1, num_users, num_heads, head_dim]
-    cache_shape = [num_users, num_heads, max_seq_len, head_dim]
-    cache = torch.randn(cache_shape).bfloat16().float()
-    cachett = ttl.tensor.Tensor(cache, cache_dtype).to(ttl.tensor.Layout.TILE).to(device)
-    x = torch.randn(input_shape).bfloat16().float()
-    x_pad = torch.nn.functional.pad(x, (0, 0, 0, 32 - num_heads), "constant", 0)
-
-    xt = ttl.tensor.Tensor(x_pad, input_dtype).to(ttl.tensor.Layout.TILE)
-    # Input is sharded
-    compute_grid_size = device.compute_with_storage_grid_size()
-    num_cores = num_users
-    shard_grid = ttl.tensor.CoreRangeSet(ttl.tensor.num_cores_to_corerange_set(num_cores, compute_grid_size, True))
-    input_shard_spec = ttl.tensor.ShardSpec(
-        shard_grid,
-        [
-            xt.volume() // xt.get_legacy_shape()[-1] // num_cores,
-            xt.get_legacy_shape()[-1],
-        ],
-        ttl.tensor.ShardOrientation.ROW_MAJOR,
-        False,
-    )
-    input_mem_config = ttl.tensor.MemoryConfig(
-        ttl.tensor.TensorMemoryLayout.HEIGHT_SHARDED, ttl.tensor.BufferType.L1, input_shard_spec
-    )
-    xt = xt.to(device, input_mem_config)
-
-    # Create arbitrary update indices
-    cache_idxs = [cache_idx + i * 17 for i in range(num_users)]
-
-    cachett = ttnn.experimental.paged_cache.paged_update_cache(cachett, xt, update_idxs=cache_idxs)
-
-    for i in range(num_users):
-        update_idx = cache_idxs[i]
-        x_view = x.permute(1, 0, 2, 3)[i, ...]
-        cache[i, 0:num_heads, update_idx : update_idx + x.shape[-2], 0 : x.shape[-1]] = x_view
-
-    tt_got_back = cachett.cpu().to(ttl.tensor.Layout.ROW_MAJOR).to_torch()
-
-    tt_updated_slice = []
-    for i in range(num_users):
-        update_idx = cache_idxs[i]
-        tt_slice = tt_got_back[i, 0:num_heads, update_idx : update_idx + x.shape[-2], 0 : x.shape[-1]]
-        tt_updated_slice.append(tt_slice)
-    tt_updated_slice = torch.stack(tt_updated_slice, dim=0).permute(1, 0, 2, 3)
-
-    if input_dtype == ttl.tensor.DataType.BFLOAT16 and cache_dtype == input_dtype:
-        eq_cache, output_cache = comp_equal(cache, tt_got_back)  # checks the entire kv cache
-        eq_update, output_update = comp_equal(x, tt_updated_slice)  # checks the updated parts
-    else:
-        eq_cache, output_cache = comp_pcc(cache, tt_got_back)  # checks the entire kv cache
-        eq_update, output_update = comp_pcc(x, tt_updated_slice)  # checks the updated parts
-    logger.debug(output_cache)
-    logger.debug(output_update)
-    assert eq_cache and eq_update
-
-
-@pytest.mark.skip("Test case covered by others")
-@pytest.mark.parametrize("check_memory", [False])
-@pytest.mark.parametrize("head_dim", [128])
-@pytest.mark.parametrize("max_seq_len", [2048])
-@pytest.mark.parametrize("num_users", [32])
-@pytest.mark.parametrize("num_heads", [1])
-@pytest.mark.parametrize("input_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-@pytest.mark.parametrize("cache_idx", [0, 1, 127, 1057])
-@pytest.mark.parametrize("cache_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-def test_update_cache_decode(
-    check_memory,
-    cache_idx,
-    head_dim,
-    max_seq_len,
-    num_users,
-    num_heads,
-    input_dtype,
-    cache_dtype,
-    device,
-    use_program_cache,
-):
-    if check_memory:
-        # Create dram tensors to check for overflow
-        cache_shape = [num_users, num_heads, max_seq_len, head_dim]
-        dram_low = ttl.tensor.Tensor(torch.zeros(cache_shape), cache_dtype).to(ttl.tensor.Layout.TILE).to(device)
-        reserved_space = ttl.tensor.Tensor(torch.zeros(cache_shape), cache_dtype).to(ttl.tensor.Layout.TILE).to(device)
-        dram_high = ttl.tensor.Tensor(torch.zeros(cache_shape), cache_dtype).to(ttl.tensor.Layout.TILE).to(device)
-        reserved_space.deallocate(True)
-
-        # Create sharded tensors to check for overflow
-        input_shape = [1, num_users, num_heads, head_dim]
-        x = torch.zeros(input_shape)
-        x_pad = torch.nn.functional.pad(x, (0, 0, 0, 32 - num_heads), "constant", 0)
-
-        xt = ttl.tensor.Tensor(x_pad, input_dtype).to(ttl.tensor.Layout.TILE)
-        # Input is sharded
-        compute_grid_size = device.compute_with_storage_grid_size()
-        num_cores = num_users
-        shard_grid = ttl.tensor.CoreRangeSet(ttl.tensor.num_cores_to_corerange_set(num_cores, compute_grid_size, True))
-        input_shard_spec = ttl.tensor.ShardSpec(
-            shard_grid,
-            [
-                xt.volume() // xt.get_legacy_shape()[-1] // num_cores,
-                xt.get_legacy_shape()[-1],
-            ],
-            ttl.tensor.ShardOrientation.ROW_MAJOR,
-            False,
-        )
-        input_mem_config = ttl.tensor.MemoryConfig(
-            ttl.tensor.TensorMemoryLayout.HEIGHT_SHARDED, ttl.tensor.BufferType.L1, input_shard_spec
-        )
-        sharded_low = xt.to(device, input_mem_config)
-        sharded_reserved = ttl.tensor.Tensor(x_pad, input_dtype).to(ttl.tensor.Layout.TILE).to(device, input_mem_config)
-        sharded_high = ttl.tensor.Tensor(x_pad, input_dtype).to(ttl.tensor.Layout.TILE).to(device, input_mem_config)
-        sharded_reserved.deallocate(True)
-
-    run_test_update_cache_decode(
-        cache_idx, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-    )
-
-    if check_memory:
-        # Check for overflow
-        def check_zero(tensor):
-            assert (tensor == 0).all()
-
-        dram_low = dram_low.cpu().to(ttl.tensor.Layout.ROW_MAJOR).to_torch()
-        dram_high = dram_high.cpu().to(ttl.tensor.Layout.ROW_MAJOR).to_torch()
-        sharded_low = sharded_low.cpu().to(ttl.tensor.Layout.ROW_MAJOR).to_torch()
-        sharded_high = sharded_high.cpu().to(ttl.tensor.Layout.ROW_MAJOR).to_torch()
-
-        check_zero(dram_low)
-        check_zero(dram_high)
-        check_zero(sharded_low)
-        check_zero(sharded_high)
-
-
-@pytest.mark.skipif(is_grayskull(), reason="Grayskull does not support paged cache")
-@pytest.mark.parametrize("head_dim", [128])
-@pytest.mark.parametrize("max_seq_len", [2048])
-@pytest.mark.parametrize("num_users", [32])
-@pytest.mark.parametrize("num_heads", [1])
-@pytest.mark.parametrize("input_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-@pytest.mark.parametrize("cache_idx", [127, 1057])
-@pytest.mark.parametrize("cache_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-def test_update_cache_decode_program_cache(
-    cache_idx,
-    head_dim,
-    max_seq_len,
-    num_users,
-    num_heads,
-    input_dtype,
-    cache_dtype,
-    device,
-    use_program_cache,
-):
-    dummy_tensors = []
-    for i in range(2):
-        # Create dram tensors to check for overflow
-        cache_shape = [num_users, num_heads, max_seq_len, head_dim]
-        dram_low = ttl.tensor.Tensor(torch.zeros(cache_shape), cache_dtype).to(ttl.tensor.Layout.TILE).to(device)
-        dummy_tensors.append(dram_low)
-
-        # Create sharded tensors to check for overflow
-        input_shape = [1, num_users, num_heads, head_dim]
-        x = torch.zeros(input_shape)
-        x_pad = torch.nn.functional.pad(x, (0, 0, 0, 32 - num_heads), "constant", 0)
-
-        xt = ttl.tensor.Tensor(x_pad, input_dtype).to(ttl.tensor.Layout.TILE)
-        # Input is sharded
-        compute_grid_size = device.compute_with_storage_grid_size()
-        num_cores = num_users
-        shard_grid = ttl.tensor.CoreRangeSet(ttl.tensor.num_cores_to_corerange_set(num_cores, compute_grid_size, True))
-        input_shard_spec = ttl.tensor.ShardSpec(
-            shard_grid,
-            [
-                xt.volume() // xt.get_legacy_shape()[-1] // num_cores,
-                xt.get_legacy_shape()[-1],
-            ],
-            ttl.tensor.ShardOrientation.ROW_MAJOR,
-            False,
-        )
-        input_mem_config = ttl.tensor.MemoryConfig(
-            ttl.tensor.TensorMemoryLayout.HEIGHT_SHARDED, ttl.tensor.BufferType.L1, input_shard_spec
-        )
-        sharded_low = xt.to(device, input_mem_config)
-        dummy_tensors.append(sharded_low)
-
-        run_test_update_cache_decode(
-            cache_idx, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-        )
-        # Test that cache_idx is correctly updated between cached runs
-        run_test_update_cache_decode(
-            cache_idx + 1, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-        )
-
-    assert device.num_program_cache_entries() == 1
-
-
-def run_test_tensor_index_update_cache_decode(
-    cache_idx, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-):
-    input_shape = [1, num_users, num_heads, head_dim]
-    cache_shape = [num_users, num_heads, max_seq_len, head_dim]
-    cache = torch.randn(cache_shape).bfloat16().float()
-
-    cachett = ttl.tensor.Tensor(cache, cache_dtype).to(ttl.tensor.Layout.TILE).to(device)
-    x = torch.randn(input_shape).bfloat16().float()
-    x_pad = torch.nn.functional.pad(x, (0, 0, 0, 32 - num_heads), "constant", 0)
-
-    xt = ttl.tensor.Tensor(x_pad, input_dtype).to(ttl.tensor.Layout.TILE)
-    # Input is sharded
-    compute_grid_size = device.compute_with_storage_grid_size()
-    num_cores = num_users
-    shard_grid = ttl.tensor.CoreRangeSet(ttl.tensor.num_cores_to_corerange_set(num_cores, compute_grid_size, True))
-    input_shard_spec = ttl.tensor.ShardSpec(
-        shard_grid,
-        [
-            xt.volume() // xt.get_legacy_shape()[-1] // num_cores,
-            xt.get_legacy_shape()[-1],
-        ],
-        ttl.tensor.ShardOrientation.ROW_MAJOR,
-        False,
-    )
-    input_mem_config = ttl.tensor.MemoryConfig(
-        ttl.tensor.TensorMemoryLayout.HEIGHT_SHARDED, ttl.tensor.BufferType.L1, input_shard_spec
-    )
-    xt = xt.to(device, input_mem_config)
-
-    # Create arbitrary update indices
-    cache_idxs = [cache_idx + i * 17 for i in range(num_users)]
-    logger.info(f"cache_idxs: {cache_idxs}")
-    cache_idxs_tt = ttl.tensor.Tensor(torch.tensor(cache_idxs), ttl.tensor.DataType.INT32).to(device)
-
-    cachett = ttnn.experimental.paged_cache.paged_update_cache(cachett, xt, update_idxs_tensor=cache_idxs_tt)
-
-    for i in range(num_users):
-        update_idx = cache_idxs[i]
-        x_view = x.permute(1, 0, 2, 3)[i, ...]
-        cache[i, 0:num_heads, update_idx : update_idx + x.shape[-2], 0 : x.shape[-1]] = x_view
-
-    tt_got_back = cachett.cpu().to(ttl.tensor.Layout.ROW_MAJOR).to_torch()
-
-    tt_updated_slice = []
-    for i in range(num_users):
-        update_idx = cache_idxs[i]
-        tt_slice = tt_got_back[i, 0:num_heads, update_idx : update_idx + x.shape[-2], 0 : x.shape[-1]]
-        tt_updated_slice.append(tt_slice)
-    tt_updated_slice = torch.stack(tt_updated_slice, dim=0).permute(1, 0, 2, 3)
-
-    if input_dtype == ttl.tensor.DataType.BFLOAT16 and cache_dtype == input_dtype:
-        eq_cache, output_cache = comp_equal(cache, tt_got_back)  # checks the entire kv cache
-        eq_update, output_update = comp_equal(x, tt_updated_slice)  # checks the updated parts
-    else:
-        eq_cache, output_cache = comp_pcc(cache, tt_got_back)  # checks the entire kv cache
-        eq_update, output_update = comp_pcc(x, tt_updated_slice)  # checks the updated parts
-    logger.debug(output_cache)
-    logger.debug(output_update)
-    assert eq_cache and eq_update
-
-
-@pytest.mark.skip("Test case covered by others")
-@pytest.mark.parametrize("head_dim", [128])
-@pytest.mark.parametrize("max_seq_len", [2048])
-@pytest.mark.parametrize("num_users", [32])
-@pytest.mark.parametrize("num_heads", [1])
-@pytest.mark.parametrize("input_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-@pytest.mark.parametrize("cache_idx", [0, 1, 127, 1057])
-@pytest.mark.parametrize("cache_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-def test_tensor_index_update_cache_decode(
-    cache_idx,
-    head_dim,
-    max_seq_len,
-    num_users,
-    num_heads,
-    input_dtype,
-    cache_dtype,
-    device,
-    use_program_cache,
-):
-    run_test_tensor_index_update_cache_decode(
-        cache_idx, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-    )
-
-
-@pytest.mark.skipif(is_grayskull(), reason="Grayskull does not support paged cache")
-@pytest.mark.parametrize("head_dim", [128])
-@pytest.mark.parametrize("max_seq_len", [2048])
-@pytest.mark.parametrize("num_users", [32])
-@pytest.mark.parametrize("num_heads", [1])
-@pytest.mark.parametrize("input_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-@pytest.mark.parametrize("cache_idx", [127, 1057])
-@pytest.mark.parametrize("cache_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-def test_tensor_index_update_cache_decode_program_cache(
-    cache_idx,
-    head_dim,
-    max_seq_len,
-    num_users,
-    num_heads,
-    input_dtype,
-    cache_dtype,
-    device,
-    use_program_cache,
-):
-    for _ in range(2):
-        run_test_tensor_index_update_cache_decode(
-            cache_idx, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-        )
-
-    assert device.num_program_cache_entries() == 1
-
-
-def run_test_paged_update_cache_decode(
-    cache_idx, block_size, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-):
-    max_num_blocks_per_seq = max_seq_len // block_size
-    assert max_num_blocks_per_seq * block_size == max_seq_len
-    max_num_blocks = num_users * max_seq_len // block_size
-    assert max_num_blocks * block_size == num_users * max_seq_len
-
-    input_shape = [1, num_users, num_heads, head_dim]
-    cache_shape = [num_users, num_heads, max_seq_len, head_dim]
-    cache = torch.randn(cache_shape).bfloat16().float()
-    paged_cache = (
-        cache.reshape(num_users, num_heads, max_num_blocks_per_seq, block_size, head_dim)
-        .transpose(1, 2)
-        .reshape(max_num_blocks, num_heads, block_size, head_dim)
-    )  # [num_users * max_num_blocks_per_seq, num_heads, block_size, head_dim]
-
-    # Shuffle paged KV cache according to some random page_table
-    permutation = torch.randperm(max_num_blocks)
-    shuffled_page_cache = paged_cache[permutation]
-    reverse_permutation = torch.argsort(permutation)
-    # page_table is the reverse permutation from shuffled -> unshuffled, and is used to map
-    # a virtual block to the physical block id.
-    page_table = reverse_permutation.reshape(num_users, max_num_blocks_per_seq)
-    # logger.info(f"page_table: {page_table}")
-    unshuffled_page_cache = shuffled_page_cache[reverse_permutation]
-
-    paged_cache_back = (
-        unshuffled_page_cache.reshape(num_users, max_num_blocks_per_seq, num_heads, block_size, head_dim)
-        .transpose(1, 2)
-        .reshape(num_users, num_heads, max_seq_len, head_dim)
-    )
-    assert torch.allclose(paged_cache_back, cache)
-
-    cachett = ttl.tensor.Tensor(shuffled_page_cache, cache_dtype).to(ttl.tensor.Layout.TILE).to(device)
-    x = torch.randn(input_shape).bfloat16().float()
-    x_pad = torch.nn.functional.pad(x, (0, 0, 0, 32 - num_heads), "constant", 0)
-
-    xt = ttl.tensor.Tensor(x_pad, input_dtype).to(ttl.tensor.Layout.TILE)
-    # Input is sharded
-    compute_grid_size = device.compute_with_storage_grid_size()
-    num_cores = num_users
-    shard_grid = ttl.tensor.CoreRangeSet(ttl.tensor.num_cores_to_corerange_set(num_cores, compute_grid_size, True))
-    input_shard_spec = ttl.tensor.ShardSpec(
-        shard_grid,
-        [
-            xt.volume() // xt.get_legacy_shape()[-1] // num_cores,
-            xt.get_legacy_shape()[-1],
-        ],
-        ttl.tensor.ShardOrientation.ROW_MAJOR,
-        False,
-    )
-    input_mem_config = ttl.tensor.MemoryConfig(
-        ttl.tensor.TensorMemoryLayout.HEIGHT_SHARDED, ttl.tensor.BufferType.L1, input_shard_spec
-    )
-    xt = xt.to(device, input_mem_config)
-
-    # Create arbitrary update indices
-    cache_idxs = [cache_idx + i * 17 for i in range(num_users)]
-    # Arbitrary user is "dropped", to test skipping in kernel
-    cache_idxs[num_users // 2] = -1
-    # logger.info(f"cache_idxs: {cache_idxs}")
-    cache_idxs_tt = ttl.tensor.Tensor(torch.tensor(cache_idxs), ttl.tensor.DataType.INT32).to(device)
-    page_table_tt = ttl.tensor.Tensor(page_table, ttl.tensor.DataType.INT32).to(device)
-
-    cachett = ttnn.experimental.paged_cache.paged_update_cache(
-        cachett, xt, update_idxs_tensor=cache_idxs_tt, page_table=page_table_tt
-    )
-
-    for i in range(num_users):
-        update_idx = cache_idxs[i]
-        if update_idx == -1:
-            continue
-        x_view = x.permute(1, 0, 2, 3)[i, ...]
-        cache[i, 0:num_heads, update_idx : update_idx + x.shape[-2], 0 : x.shape[-1]] = x_view
-
-    # Unshuffle paged cache and review it as unpaged cache
-    tt_got_back_shuffled = cachett.cpu().to(ttl.tensor.Layout.ROW_MAJOR).to_torch()
-    tt_got_back_unshuffled = tt_got_back_shuffled[reverse_permutation]
-    tt_got_back = (
-        tt_got_back_unshuffled.reshape(num_users, max_num_blocks_per_seq, num_heads, block_size, head_dim)
-        .transpose(1, 2)
-        .reshape(num_users, num_heads, max_seq_len, head_dim)
-    )
-
-    tt_updated_slice = []
-    for i in range(num_users):
-        update_idx = cache_idxs[i]
-        if update_idx == -1:
-            # Skipped users should compare to the original cache
-            update_idx = 0
-            x[:, i : i + 1, :, :] = cache[
-                i : i + 1, 0:num_heads, update_idx : update_idx + x.shape[-2], 0 : x.shape[-1]
-            ]
-        tt_slice = tt_got_back[i, 0:num_heads, update_idx : update_idx + x.shape[-2], 0 : x.shape[-1]]
-        tt_updated_slice.append(tt_slice)
-    tt_updated_slice = torch.stack(tt_updated_slice, dim=0).permute(1, 0, 2, 3)
-
-    if input_dtype == ttl.tensor.DataType.BFLOAT16 and cache_dtype == input_dtype:
-        eq_cache, output_cache = comp_equal(cache, tt_got_back)  # checks the entire kv cache
-        eq_update, output_update = comp_equal(x, tt_updated_slice)  # checks the updated parts
-    else:
-        eq_cache, output_cache = comp_pcc(cache, tt_got_back)  # checks the entire kv cache
-        eq_update, output_update = comp_pcc(x, tt_updated_slice)  # checks the updated parts
-    logger.debug(output_cache)
-    logger.debug(output_update)
-    assert eq_cache and eq_update
-
-
-@pytest.mark.skip("Test case covered by others")
-@pytest.mark.parametrize("block_size", [64, 128], ids=["block64", "block128"])
-@pytest.mark.parametrize("head_dim", [128])
-@pytest.mark.parametrize("max_seq_len", [2048])
-@pytest.mark.parametrize("num_users", [32])
-@pytest.mark.parametrize("num_heads", [1])
-@pytest.mark.parametrize("input_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-@pytest.mark.parametrize("cache_idx", [0, 1, 127, 1057])
-@pytest.mark.parametrize("cache_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-def test_paged_update_cache_decode(
-    cache_idx,
-    block_size,
-    head_dim,
-    max_seq_len,
-    num_users,
-    num_heads,
-    input_dtype,
-    cache_dtype,
-    device,
-    use_program_cache,
-):
-    run_test_paged_update_cache_decode(
-        cache_idx, block_size, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-    )
-
-
-@pytest.mark.skipif(is_grayskull(), reason="Grayskull does not support paged cache")
-@pytest.mark.parametrize("block_size", [64, 128], ids=["block64", "block128"])
-@pytest.mark.parametrize("head_dim", [128])
-@pytest.mark.parametrize("max_seq_len", [2048])
-@pytest.mark.parametrize("num_users", [32])
-@pytest.mark.parametrize("num_heads", [1])
-@pytest.mark.parametrize("input_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-@pytest.mark.parametrize("cache_idx", [127, 1057])
-@pytest.mark.parametrize("cache_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-def test_paged_update_cache_decode_program_caching(
-    cache_idx,
-    block_size,
-    head_dim,
-    max_seq_len,
-    num_users,
-    num_heads,
-    input_dtype,
-    cache_dtype,
-    device,
-    use_program_cache,
-):
-    dummy_tensors = []
-    for i in range(2):
-        # Create dram tensors to check for overflow
-        cache_shape = [num_users, num_heads, max_seq_len, head_dim]
-        dram_low = ttl.tensor.Tensor(torch.zeros(cache_shape), cache_dtype).to(ttl.tensor.Layout.TILE).to(device)
-        dummy_tensors.append(dram_low)
-
-        # Create sharded tensors to check for overflow
-        input_shape = [1, num_users, num_heads, head_dim]
-        x = torch.zeros(input_shape)
-        x_pad = torch.nn.functional.pad(x, (0, 0, 0, 32 - num_heads), "constant", 0)
-
-        xt = ttl.tensor.Tensor(x_pad, input_dtype).to(ttl.tensor.Layout.TILE)
-        # Input is sharded
-        compute_grid_size = device.compute_with_storage_grid_size()
-        num_cores = num_users
-        shard_grid = ttl.tensor.CoreRangeSet(ttl.tensor.num_cores_to_corerange_set(num_cores, compute_grid_size, True))
-        input_shard_spec = ttl.tensor.ShardSpec(
-            shard_grid,
-            [
-                xt.volume() // xt.get_legacy_shape()[-1] // num_cores,
-                xt.get_legacy_shape()[-1],
-            ],
-            ttl.tensor.ShardOrientation.ROW_MAJOR,
-            False,
-        )
-        input_mem_config = ttl.tensor.MemoryConfig(
-            ttl.tensor.TensorMemoryLayout.HEIGHT_SHARDED, ttl.tensor.BufferType.L1, input_shard_spec
-        )
-        sharded_low = xt.to(device, input_mem_config)
-        dummy_tensors.append(sharded_low)
-
-        run_test_paged_update_cache_decode(
-            cache_idx, block_size, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-        )
-
-        run_test_paged_update_cache_decode(
-            cache_idx + 10, block_size, head_dim, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-        )
-
-    assert device.num_program_cache_entries() == 1
-
-
-def run_test_paged_fill_cache(
-    block_size, head_dim, user_seq_len, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-):
-    max_num_blocks_per_seq = max_seq_len // block_size
-    assert max_num_blocks_per_seq * block_size == max_seq_len
-    max_num_blocks = num_users * max_seq_len // block_size
-    assert max_num_blocks * block_size == num_users * max_seq_len
-
-    input_shape = [1, num_heads, user_seq_len, head_dim]
-    cache_shape = [num_users, num_heads, max_seq_len, head_dim]
-    cache = torch.randn(cache_shape).bfloat16().float()
-
-    # Turn cache into paged cache
-    paged_cache = (
-        cache.reshape(num_users, num_heads, max_num_blocks_per_seq, block_size, head_dim)
-        .transpose(1, 2)
-        .reshape(max_num_blocks, num_heads, block_size, head_dim)
-    )  # [num_users * max_num_blocks_per_seq, num_heads, block_size, head_dim]
-
-    # Shuffle paged KV cache according to some random page_table
-    permutation = torch.randperm(max_num_blocks)
-    shuffled_page_cache = paged_cache[permutation]
-    reverse_permutation = torch.argsort(permutation)
-    # page_table is the reverse permutation from shuffled -> unshuffled, and is used to map
-    # a virtual block to the physical block id.
-    page_table = reverse_permutation.reshape(num_users, max_num_blocks_per_seq)
-
-    # logger.info(f"page_table: {page_table}")
-    unshuffled_page_cache = shuffled_page_cache[reverse_permutation]
-
-    paged_cache_back = (
-        unshuffled_page_cache.reshape(num_users, max_num_blocks_per_seq, num_heads, block_size, head_dim)
-        .transpose(1, 2)
-        .reshape(num_users, num_heads, max_seq_len, head_dim)
-    )
-    # Check that we can convert from normal to paged to normal
-    assert torch.allclose(paged_cache_back, cache)
-
-    cachett = ttl.tensor.Tensor(shuffled_page_cache, cache_dtype).to(ttl.tensor.Layout.TILE).to(device)
-    page_table_tt = ttl.tensor.Tensor(page_table, ttl.tensor.DataType.INT32).to(device)
-
-    # Update cache for every user
-    for i in range(num_users):
-        x = torch.randn(input_shape).bfloat16().float()
-        xt = ttl.tensor.Tensor(x, input_dtype).to(ttl.tensor.Layout.TILE).to(device)
-
-        cachett = ttnn.experimental.paged_cache.paged_fill_cache(cachett, xt, page_table_tt, batch_idx=i)
-        cache[i : i + 1, :, : x.shape[-2], :] = x
-
-    # Unshuffle paged cache and review it as unpaged cache
-    tt_got_back_shuffled = cachett.cpu().to(ttl.tensor.Layout.ROW_MAJOR).to_torch()
-    tt_got_back_unshuffled = tt_got_back_shuffled[reverse_permutation]
-    tt_got_back = (
-        tt_got_back_unshuffled.reshape(num_users, max_num_blocks_per_seq, num_heads, block_size, head_dim)
-        .transpose(1, 2)
-        .reshape(num_users, num_heads, max_seq_len, head_dim)
-    )
-
-    if input_dtype == ttl.tensor.DataType.BFLOAT16 and cache_dtype == input_dtype:
-        eq, output = comp_equal(cache, tt_got_back)
-    else:
-        eq, output = comp_pcc(cache, tt_got_back)
-    logger.info(output)
-    assert eq
-
-
-@pytest.mark.skip("Test case covered by others")
-@pytest.mark.parametrize("block_size", [64, 128], ids=["block64", "block128"])
-@pytest.mark.parametrize("head_dim", [128])
-@pytest.mark.parametrize("user_seq_len", [128, 160, 1984, 2048])
-@pytest.mark.parametrize("max_seq_len", [2048])
-@pytest.mark.parametrize("num_users", [32])
-@pytest.mark.parametrize("num_heads", [1])
-@pytest.mark.parametrize("input_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-def test_paged_fill_cache(
-    block_size,
-    head_dim,
-    user_seq_len,
-    max_seq_len,
-    num_users,
-    num_heads,
-    input_dtype,
-    device,  # use_program_cache
-):
-    logger.warning("Forcing cache_dtype to be same as input_dtype. Change test case when this is not required")
-    cache_dtype = input_dtype
-    run_test_paged_fill_cache(
-        block_size, head_dim, user_seq_len, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-    )
-
-
-@pytest.mark.skipif(is_grayskull(), reason="Grayskull does not support paged cache")
-@pytest.mark.parametrize("block_size", [64, 128], ids=["block64", "block128"])
-@pytest.mark.parametrize("head_dim", [128])
-@pytest.mark.parametrize("user_seq_len", [128, 160, 1984, 2048])
-@pytest.mark.parametrize("max_seq_len", [2048])
-@pytest.mark.parametrize("num_users", [32])
-@pytest.mark.parametrize("num_heads", [1])
-@pytest.mark.parametrize("input_dtype", [ttl.tensor.DataType.BFLOAT16, ttl.tensor.DataType.BFLOAT8_B])
-def test_paged_fill_cache_program_cache(
-    block_size, head_dim, user_seq_len, max_seq_len, num_users, num_heads, input_dtype, device, use_program_cache
-):
-    cache_dtype = input_dtype
-    run_test_paged_fill_cache(
-        block_size, head_dim, user_seq_len, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-    )
-    run_test_paged_fill_cache(
-        block_size, head_dim, user_seq_len, max_seq_len, num_users, num_heads, input_dtype, cache_dtype, device
-    )
-
-    assert device.num_program_cache_entries() == 1
