+++ ttnn/cpp/ttnn/deprecated/tt_dnn/op_library/CMakeLists.txt
@@ -168,6 +168,8 @@ set(TT_DNN_SRCS
     ${CMAKE_CURRENT_SOURCE_DIR}/embeddings/embeddings_op.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/update_cache/multi_core/update_cache_op_multi_core.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/update_cache/update_cache_op.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/paged_update_cache/multi_core/paged_update_cache_op_multi_core.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/paged_update_cache/paged_update_cache_op.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/fold/fold_op.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/fold/single_core/fold_op_single_core.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/fold/multi_core/fold_op_multi_core.cpp
