+++ ttnn/cpp/ttnn/deprecated/tt_dnn/op_library/cb_utils.hpp
@@ -2,8 +2,6 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-#pragma once
-
 #include "tt_metal/host_api.hpp"
 
 namespace tt::tt_metal {
