+++ ttnn/cpp/ttnn/deprecated/tt_lib/csrc/operations/primary/transformers/module.hpp
@@ -6,6 +6,7 @@
 
 #include "ttnn/deprecated/tt_dnn/op_library/transformer_tms/transformer_tms.hpp"
 #include "ttnn/deprecated/tt_dnn/op_library/sdpa/sdpa_op.hpp"
+#include "ttnn/deprecated/tt_dnn/op_library/paged_update_cache/paged_update_cache_op.hpp"
 
 #include <pybind11/pybind11.h>
 #include <pybind11/stl.h>
@@ -123,6 +124,20 @@ void py_module(py::module& m_transformers) {
 
         "Accepts a `SDPAMultiCoreProgramConfig` which specifies the grid size and chunk tiles in the K/V/Mask sequence lengths (Q chunk tiles is not used). The op parallelizes over `b` and K/V/Mask's `s` dimension."
         );
+
+    m_transformers.def(
+        "paged_update_cache",
+        &paged_update_cache,
+        py::arg("cache_tensor").noconvert(),
+        py::arg("input_tensor").noconvert(),
+        py::arg("update_idxs").noconvert(),
+        py::arg("batch_offset") = 0,
+        py::arg("compute_kernel_config").noconvert() = std::nullopt,
+        R"doc(
+        Paged update cache operation. This operation expects the following inputs: cache_tensor of shape [B, 1, kv_len, head_dim] and input_tensor of shape [1, B, 1[32], head_dim] where input_tensor is height sharded on B cores. update_idxs will specify for each batch element which token to update in the cache.
+        )doc"
+    );
+
 
 }
