+++ ttnn/cpp/ttnn/operations/experimental/experimental_pybind.hpp
@@ -8,16 +8,12 @@
 #include <pybind11/stl.h>
 
 #include "ttnn/operations/experimental/transformer/transformer_pybind.hpp"
-#include "ttnn/operations/experimental/paged_cache/paged_cache_pybind.hpp"
 
 namespace ttnn::operations::experimental {
 
 void py_module(py::module& module) {
     // Transformer ops
     transformer::detail::bind_experimental_transformer_operations(module);
-
-    // Paged update cache ops
-    paged_cache::detail::bind_experimental_paged_cache_operations(module);
 }
 
 }  // namespace ttnn::operations::experimental
